Moved client files for submission
