+++ packages/component-submission/client/ProgressBar.js
