+++ packages/component-submission/client/steps/Author/AuthorPage.md
