+++ packages/component-submission/client/steps/Author/schema.js
