+++ packages/component-submission/client/steps/Disclosure/schema.js
