+++ packages/component-submission/client/steps/Editors/EditorsPage.js
