+++ packages/component-submission/client/steps/Editors/schema.js
