+++ packages/component-submission/client/steps/Files/CoverLetterEditor.js
