+++ packages/component-submission/client/steps/Files/ManuscriptUpload.js
