+++ packages/component-submission/client/steps/Files/ManuscriptUpload.md
