+++ packages/component-submission/client/steps/Files/ManuscriptUpload.test.js
