+++ packages/component-submission/client/steps/Files/index.test.js
