+++ packages/component-submission/client/steps/Files/schema.js
