+++ packages/component-submission/client/steps/Files/utils.js
